requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.1.4
djangorestframework==3.15.2
idna==3.10
psycopg2-binary==2.9.10
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
annotated-types==0.7.0
attrs==25.1.0
confluent-kafka==2.4.0
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
logging==0.4.9.6
orjson==3.10.15
pydantic==2.10.6
pydantic-settings==2.6.1
pydantic_core==2.27.2
quixstreams==3.7.0
referencing==0.36.2
rocksdict==0.3.25
rpds-py==0.22.3
typing_extensions==4.12.2
